fix(test): type adjustments and standardization
