+++ .husky/pre-commit
@@ -1,1 +1,1 @@
-npx lint-staged
+npm run format && npx lint-staged
