+++ package.json
@@ -34,10 +34,8 @@
   "dependencies": {
     "cheerio": "^1.0.0-rc.12",
     "imapflow": "^1.0.152",
-    "jest": "^29.7.0",
     "mailparser": "^3.6.7",
-    "nodemailer": "^6.9.10",
-    "repeat-string": "^1.6.1"
+    "nodemailer": "^6.9.10"
   },
   "devDependencies": {
     "@size-limit/preset-small-lib": "^11.0.2",
@@ -46,8 +44,10 @@
     "@types/mailparser": "^3.4.4",
     "@types/nodemailer": "^6.4.14",
     "husky": "^9.0.11",
+    "jest": "^29.7.0",
     "lint-staged": "^15.2.2",
     "prettier": "3.2.5",
+    "repeat-string": "^1.6.1",
     "size-limit": "^11.0.2",
     "tsdx": "^0.14.1",
     "tslib": "^2.6.2",
