+++ test/data.ts
@@ -1,0 +1,5 @@
+export const testUser = 'pasquale.corkery@ethereal.email';
+export const testPass = 'ew7YV7AGDmvD9T3yBs';
+
+export const regexEmail =
+  /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/g;
